turned graphKernelizer into instantiable class
